refactor(v2): bump module path to /v2, drop v1 compat wrappers, promote errs to public subpackage
Phase 4 of the v2 restructure — the actual breaking change. Master
is now the v2 line; v1.x users on the old import path stay supported
via the release/v1.x branch already cut from v1.4.1 in Phase 0 PR-B.

Module path:
- go.mod: github.com/hishamkaram/gismanager → github.com/hishamkaram/gismanager/v2
- Every internal import bumped via sed (cmd/, convert/, publish/,
  internal/, examples/convert_pipeline)

Public API surface in v2:

- github.com/hishamkaram/gismanager/v2/publish: Manager, Layer,
  walk + load + publish pipeline
- github.com/hishamkaram/gismanager/v2/convert: stateless GDAL
  CLI-equivalents + ToPMTiles
- github.com/hishamkaram/gismanager/v2/errs: PROMOTED from
  internal/errs — the *GISError envelope and Err* sentinels are
  now part of the public surface so external callers can match
  via errors.Is / errors.As against the canonical sentinels
  rather than v1's gismanager.Err* aliases (which no longer exist)

Compat layer removed:
- compat_convert.go: deleted
- compat_publish.go: deleted
- errors.go (root alias shim): deleted
- errors_test.go (alias-chain test): deleted
- log.go (root GetLogger wrapper): deleted
- log_test.go: deleted

Root gismanager package is now a doc-only marker. doc.go rewritten
to point users at the three v2 subpackages (publish, convert, errs)
and document the v1 → v2 migration story (release/v1.x for stayers,
MIGRATING.md for migrators).

internal/slogx remains internal — v2 callers construct their own
*slog.Logger and pass it via WithLogger; the project no longer
exposes a default-logger constructor.

Build-tooling updates:
- Dockerfile ldflags package paths bumped to /v2/cmd/internal/cli
- Makefile LDFLAGS_PKG bumped to /v2/cmd/internal/cli
- examples/otel_pipeline/go.mod: module path bumped to /v2/...,
  require + replace target updated to gismanager/v2 v2.0.0-...,
  go mod tidy refreshed transitive deps
- examples/otel_pipeline/main.go: imports gismanager/v2/publish
  directly (no more deprecated gismanager.New wrapper)

Tests: full unit suite + integration suite (GeoServer 2.28.0,
~9s publish + ~3.5s convert) green inside the dev container. Test
output now reports under github.com/hishamkaram/gismanager/v2/...
package paths.

After this lands, Phase 5 cleans up the API surface (drop typo'd
GetGeomtryName, the deprecated GetFeatures, OTel-conventional log
keys), Phase 6 adds publish.WithPublishConcurrency, Phase 7 ships
README + MIGRATING.md + the [2.0.0] CHANGELOG stanza, and Phase 8
cuts v2.0.0 proper.
