+++ cmd/layerSchema/layerSchema.go
@@ -19,8 +19,8 @@ import (
 	"log/slog"
 	"os"
 
-	"github.com/hishamkaram/gismanager/cmd/internal/cli"
-	"github.com/hishamkaram/gismanager/publish"
+	"github.com/hishamkaram/gismanager/v2/cmd/internal/cli"
+	"github.com/hishamkaram/gismanager/v2/publish"
 )
 
 // layerEntry is the JSON shape for one yielded layer when -json is set.
