+++ convert/buildvrt_integration_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/lukeroth/gdal"
 
-	"github.com/hishamkaram/gismanager/convert"
+	"github.com/hishamkaram/gismanager/v2/convert"
 )
 
 // TestBuildVRT_TwoGeoTIFFs_Integration assembles two GeoTIFF inputs into
