+++ convert/example_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/lukeroth/gdal"
 
-	"github.com/hishamkaram/gismanager/convert"
+	"github.com/hishamkaram/gismanager/v2/convert"
 )
 
 // ExampleConvertVector demonstrates a vector format conversion from
