+++ convert/rasterize_integration_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/lukeroth/gdal"
 
-	"github.com/hishamkaram/gismanager/convert"
+	"github.com/hishamkaram/gismanager/v2/convert"
 )
 
 // TestRasterize_PolygonsToGeoTIFF_Integration burns the Africa subset
