+++ convert/vector_vsimem_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/lukeroth/gdal"
 
-	"github.com/hishamkaram/gismanager/convert"
+	"github.com/hishamkaram/gismanager/v2/convert"
 )
 
 // TestConvertVector_VsiMemDestination_Unit confirms ConvertVector accepts
