+++ publish/driver_table_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/lukeroth/gdal"
 
-	"github.com/hishamkaram/gismanager/internal/errs"
+	"github.com/hishamkaram/gismanager/v2/errs"
 )
 
 // TestGetDriver_Table is the table-driven canonical test for the
